backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy[asyncio]>=2.0
aiomysql>=0.2.0
pymysql>=1.0
pydantic>=2.5
pydantic-settings>=2.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart>=0.0.9
loguru>=0.7.0
email-validator>=2.1.0
openpyxl>=3.1
slowapi>=0.1.9
alembic>=1.13
